first pass at reorganziing singularity-eos
