+++ .gitignore
@@ -1,3 +1,5 @@
+bin
 .spack*
 spack.lock
+utils/variant
 *__pycache__*
