+++ sesame2spiner/generate_files.cpp
@@ -18,19 +18,19 @@
 #include <cmath>
 #include <cstdlib>
 
-#include "hdf5.h"
-#include "hdf5_hl.h"
-#include "nlohmann/json.hpp"
+#include <hdf5.h>
+#include <hdf5_hl.h>
+#include <nlohmann/json.hpp>
 
 #ifndef SPINER_USE_HDF
-#define SPINER_USE_HDF (1) // Enable HDF5
+#error "HDF5 must be enabled"
 #endif // SPINER_USE_HDF
 
-#include "../spiner/ports-of-call/portability.hpp"
-#include "../sp5/singularity_eos_sp5.hpp"
-#include "../spiner/databox.hpp"
-#include "../spiner/interpolation.hpp"
-#include "../spiner/sp5.hpp"
+#include <spiner/ports-of-call/portability.hpp>
+#include <sp5/singularity_eos_sp5.hpp>
+#include <spiner/databox.hpp>
+#include <spiner/interpolation.hpp>
+#include <spiner/sp5.hpp>
 
 #include "io_eospac.hpp"
 #include "generate_files.hpp"
