+++ sesame2spiner/generate_files.hpp
@@ -19,9 +19,9 @@
 
 #include <string>
 
-#include "hdf5.h"
-#include "hdf5_hl.h"
-#include "nlohmann/json.hpp"
+#include <hdf5.h>
+#include <hdf5_hl.h>
+#include <nlohmann/json.hpp>
 
 #include "io_eospac.hpp"
