+++ sesame2spiner/io_eospac.hpp
@@ -26,14 +26,13 @@
 #include <eos_Interface.h> // eospac API
 
 #ifndef SPINER_USE_HDF
-#define SPINER_USE_HDF (1) // Enable HDF5
+#error "HDF5 must be enabled"
 #endif
-#include "databox.hpp"
-#include "interpolation.hpp"
-#include "../spiner/ports-of-call/portability.hpp"
-#include "../spiner/databox.hpp"
-#include "../spiner/interpolation.hpp"
-#include "../fast-math/logs.hpp"
+
+#include <spiner/ports-of-call/portability.hpp>
+#include <spiner/databox.hpp>
+#include <spiner/interpolation.hpp>
+#include <fast-math/logs.hpp>
 
 using Spiner::DataBox;
 using Spiner::RegularGrid1D;
