+++ sesame2spiner/test.cpp
@@ -18,24 +18,24 @@
 #include <string>
 #include <iostream>
 
-#include "hdf5.h"
-#include "hdf5_hl.h"
+#include <hdf5.h>
+#include <hdf5_hl.h>
 
 #define CATCH_CONFIG_MAIN
 #include "catch.hpp"
 
-#include "../nlohmann/json.hpp"
+#include <nlohmann/json.hpp>
 
 #ifndef SPINER_USE_HDF
-#define SPINER_USE_HDF (1)
+#error "HDF5 must be enabled"
 #endif
 
-#include "../sp5/singularity_eos_sp5.hpp"
-#include "../spiner/spiner_types.hpp"
-#include "../spiner/ports-of-call/portability.hpp"
-#include "../spiner/databox.hpp"
-#include "../spiner/interpolation.hpp"
-#include "../spiner/sp5.hpp"
+#include <sp5/singularity_eos_sp5.hpp>
+#include <spiner/spiner_types.hpp>
+#include <spiner/ports-of-call/portability.hpp>
+#include <spiner/databox.hpp>
+#include <spiner/interpolation.hpp>
+#include <spiner/sp5.hpp>
 
 #include "io_eospac.hpp"
 #include "generate_files.hpp"
