+++ singularity-eos/eos/CMakeLists.txt
