+++ singularity-eos/eos/eos_gruneisen.cpp
@@ -12,7 +12,7 @@
 // publicly and display publicly, and to permit others to do so.
 //------------------------------------------------------------------------------
 
-#include "eos.hpp"
+#include <singularity-eos/eos/eos.hpp>
 
 namespace singularity {
