+++ singularity-eos/eos/eos_ideal.cpp
@@ -12,7 +12,7 @@
 // publicly and display publicly, and to permit others to do so.
 //------------------------------------------------------------------------------
 
-#include "eos.hpp"
+#include <singularity-eos/eos/eos.hpp>
 
 #define MYMAX(a,b) a > b ? a : b
