+++ singularity-eos/eos/singularity_eos.f90
