+++ test/test_pte_flag.cpp
@@ -19,10 +19,10 @@
 #include <time.h>
 #include <chrono>
 
-#include "../utils/ports-of-call/portability.hpp"
-#include "../eos/eos.hpp"
-#include "../eos/bedroom_door.hpp"
-#include "../closure/mixed_cell_models.hpp"
+#include <ports-of-call/portability.hpp>
+#include <singularity-eos/eos/eos.hpp>
+#include <singularity-eos/eos/bedroom_door.hpp>
+#include <singularity-eos/closure/mixed_cell_models.hpp>
 
 static constexpr const int NMAT {3};
 static constexpr const int NTRIAL {3};
